Refactor current page plugin
