+++ packages/rpv-current-page/src/StoreProps.ts
@@ -1,0 +1,5 @@
+interface StoreProps {
+    currentPage?: number;
+}
+
+export default StoreProps;
